agent/pyproject.toml
[project]
name = "voxarena-agent"
version = "0.1.0"
description = "VoxArena Voice Agent with Deepgram STT, Gemini LLM, and Resemble AI TTS"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents==0.12.1",
    "livekit-plugins-deepgram==0.6.0",
    "livekit-plugins-silero==0.7.0",
    "livekit-plugins-google==0.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[tool.setuptools]
py-modules = ["agent", "resemble_tts"]
